# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EvalAI API",
    "version": "1.0.0",
    "description": "AI Evaluation Platform API. See docs/api-contract.md for stability commitments."
  },
  "servers": [
    {
      "url": "https://v0-ai-evaluation-platform-nu.vercel.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/reports": {
      "post": {
        "summary": "Create signed report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["evaluationId", "evaluationRunId"],
                "properties": {
                  "evaluationId": {
                    "type": "integer"
                  },
                  "evaluationRunId": {
                    "type": "integer"
                  },
                  "expiresInDays": {
                    "type": "number"
                  },
                  "policyName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "shareToken": {
                      "type": "string"
                    },
                    "shareUrl": {
                      "type": "string"
                    },
                    "apiUrl": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/workflows": {
      "get": {
        "summary": "workflows",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/traces": {
      "get": {
        "summary": "traces",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/traces/{id}/spans": {
      "get": {
        "summary": "spans",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/api/report-cards": {
      "get": {
        "summary": "report-cards",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key or Session Token"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "VALIDATION_ERROR",
                  "RATE_LIMITED",
                  "CONFLICT",
                  "INTERNAL_ERROR",
                  "SERVICE_UNAVAILABLE",
                  "QUOTA_EXCEEDED",
                  "NO_ORG_MEMBERSHIP"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  }
}